+++ config/webhooks.php
@@ -1,0 +1,8 @@
+<?php
+
+return [
+    'enabled' => true,
+    'timeout' => 5,
+    'max_retries' => 2,
+    'parallel_dispatch' => true,
+];